Created staff. Enhanced hand system.
Players can now double tap the E key to swap what is in their left and right hands. with an accompanying animation. The animation for picking up is improved. There is now a staff near the player spawn. As of now the staff can freeze other pickups, jump pads, and snow balls.